SCUMM: Give Loom and Indy 3 their own Mac GUI header files
