+++ engines/scumm/macgui/macgui_loom.h
@@ -19,61 +19,48 @@
  *
  */
 
-#ifndef SCUMM_MACGUI_MACGUI_H
-#define SCUMM_MACGUI_MACGUI_H
+#ifndef SCUMM_MACGUI_MACGUI_LOOM_H
+#define SCUMM_MACGUI_MACGUI_LOOM_H
 
-#include "common/events.h"
+#include "common/rect.h"
 #include "common/str.h"
-
-namespace Graphics {
-class Font;
-struct Surface;
-}
 
 namespace Scumm {
 
-class ScummEngine;
-class Actor;
 class MacGuiImpl;
 
-class MacGui {
-private:
-	MacGuiImpl *_impl = nullptr;
-
+class MacLoomGui : public MacGuiImpl {
 public:
-	MacGui(ScummEngine *vm, Common::String resourceFile);
-	~MacGui();
+	MacLoomGui(ScummEngine *vm, Common::String resourceFile);
+	~MacLoomGui();
 
-	void initialize();
-	void reset();
-	void update(int delta);
-	void updateWindowManager();
+	const Common::String name() const { return "Loom"; }
 
-	void resetAfterLoad();
 	bool handleEvent(Common::Event event);
-
-	void setupCursor(int &width, int &height, int &hotspotX, int &hotspotY, int &animate);
-
-	void setPalette(const byte *palette, uint size);
 
 	const Graphics::Font *getFontByScummId(int32 id);
 
-	void drawBanner(char *message);
-	void undrawBanner();
+	void setupCursor(int &width, int &height, int &hotspotX, int &hotspotY, int &animate);
 
-	bool runQuitDialog();
-	bool runRestartDialog();
+	void resetAfterLoad();
+	void update(int delta);
 
-	// Indiana Jones and the Last Crusade
-	bool isVerbGuiActive() const;
-
-	Graphics::Surface *textArea() const;
-	void clearTextArea();
-	void initTextAreaForActor(Actor *a, byte color);
-	void printCharToTextArea(int chr, int x, int y, int color);
-
-	// Loom
 	void runDraftsInventory();
+
+protected:
+	bool getFontParams(FontId fontId, int &id, int &size, int &slant) const;
+
+	bool handleMenu(int id, Common::String &name);
+
+	void runAboutDialog();
+	bool runOpenDialog(int &saveSlotToHandle);
+	bool runSaveDialog(int &saveSlotToHandle, Common::String &name);
+	bool runOptionsDialog();
+
+private:
+	Graphics::Surface *_practiceBox = nullptr;
+	Common::Point _practiceBoxPos;
+	int _practiceBoxNotes;
 };
 
 } // End of namespace Scumm
